Add cloud optical calculations for use in TUV-x (#167)
Adds the calculation of cloud optical properties to TUV-x for photolysis
rate constant calculations. The algorithm is adapted from a
post-calculation modification of photolysis rate constants calculated by
the lookup version of TUV-x in CAM to account for the effects of clouds,
[here](https://github.com/ESCOMP/CAM/blob/47dd341f83afc1e1faf28fd4cb6e463629db4c19/src/chemistry/mozart/mo_photo.F90#L949-L965)

closes #100 

Adds a few more Valgrind suppressions for minor memory leaks during
initialization functions in 2 tests (9 and 72 bytes). We will look into
these, but they are expected to go away as TUV-x is moved out of
Fortran.