Solar zenith angle and Earth-Sun distance (#171)
Adds the calculation of solar zenith angle and Earth-Sun distance needed
as inputs to TUV-x.

closes #163 
closes #164 

Both calculations required functions that are currently in shared code
in CAM
[here](https://github.com/ESCOMP/CESM_share/blob/f6f31fd61cb8f80aee97311fcca64b3e26b0202c/src/shr_orb_mod.F90#L736-L814)
and
[here](https://github.com/ESCOMP/CESM_share/blob/f6f31fd61cb8f80aee97311fcca64b3e26b0202c/src/shr_orb_mod.F90#L110-L159)
that I put in a modified form in the `to_be_ccppized/` folder. @nusbaume
@peverwhee - if there is a better way to handle these functions, let me
know and I can update the PR.

Requires an update to CAM-SIMA that is in review to make certain
variables available through CCPP
(https://github.com/ESCOMP/CAM-SIMA/pull/325)

After more discussion, the plan is to use the shared functions in
CAM-SIMA and pass the outputs through CCPP. So, this PR will be put in
draft until this issue is finished:
https://github.com/ESCOMP/CAM-SIMA/issues/328

A related draft PR to CAM-SIMA has been added
(https://github.com/ESCOMP/CAM-SIMA/pull/337) and will be un-drafted
once this PR is merged in.