Draft aerosol classes (#209)
Adds an initial implementation of the abstract aerosol classes from CAM
along with a stub aerosol model.

closes #205 

I'll leave this in draft while we iterate on the design. This makes a
few modifications to the CAM aerosol classes' virtual functions
including:
- returning optical properties for all grid cells and wavelength bins
- having the aerosol model state own its own data, as opposed to it
being managed externally
- specifying a wavelength grid for optical properties to be returned on

This design should also support concurrent use of multiple aerosol
models and/or multiple state objects for a single aerosol model
instance.

The best place to start reviewing this is probably in
`test/musica/aerosol/mock_host.F90`, which shows how a simple host model
might use the aerosol classes with a radiation package.

@gold2718 - let me know if you think this structure will work for your
use case. Feel free to request changes or submit PRs to my dev branch.
We could also set up a meeting to discuss needs in more detail.

Is this an answer-changing PR? If so, is it a new physics package,
algorithm change, tuning change, etc?
no

If yes to the above question, describe how this code was validated with
the new/modified features:
n/a

---------

Co-authored-by: Jiwon Gim <55209567+boulderdaze@users.noreply.github.com>